refactor: move remaining `zeebe-test` utilities to elasticsearch module
