+++ exporters/elasticsearch-exporter/src/test/java/io/camunda/zeebe/exporter/AbstractElasticsearchExporterIntegrationTestCase.java
@@ -17,9 +17,9 @@ import io.camunda.zeebe.exporter.dto.GetSettingsForIndicesResponse;
 import io.camunda.zeebe.exporter.dto.GetSettingsForIndicesResponse.IndexSettings;
 import io.camunda.zeebe.exporter.util.ElasticsearchContainer;
 import io.camunda.zeebe.exporter.util.ElasticsearchNode;
+import io.camunda.zeebe.exporter.util.it.ExporterIntegrationRule;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.ValueType;
-import io.camunda.zeebe.test.exporter.ExporterIntegrationRule;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
 import java.io.IOException;
 import java.util.Map;
