+++ exporters/elasticsearch-exporter/src/test/java/io/camunda/zeebe/exporter/util/it/ExporterTestHarness.java
@@ -5,14 +5,14 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.test.exporter;
+package io.camunda.zeebe.exporter.util.it;
 
 import io.camunda.zeebe.exporter.api.Exporter;
 import io.camunda.zeebe.exporter.api.context.Configuration;
+import io.camunda.zeebe.exporter.util.record.MockRecord;
+import io.camunda.zeebe.exporter.util.record.MockRecordMetadata;
+import io.camunda.zeebe.exporter.util.record.MockRecordStream;
 import io.camunda.zeebe.protocol.record.Record;
-import io.camunda.zeebe.test.exporter.record.MockRecord;
-import io.camunda.zeebe.test.exporter.record.MockRecordMetadata;
-import io.camunda.zeebe.test.exporter.record.MockRecordStream;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
